Implement Drawable interface in Player class
The Player class now implements the Drawable interface. This change is made to add more flexibility and control over the way Player objects are drawn on the screen. Also, a redundant NullPointerException declaration in the 'update' method comment was removed as it was not needed. A new Drawable.java file is also added as part of this commit.