Refactor to use ArrayList for game objects
In order to handle multiple drawable objects, the GamePanel class has been refactored to use an ArrayList named "drawables". This replaces the single Player instance, allowing the program to handle an indefinite number of objects. This increases scalability and generalizes the approach to rendering, as all GameObjects are now stored in the "drawables" ArrayList and are handled within a loop for updating and drawing.