Remove unnecessary files, change the map and movement
These files were identified as macOS system files that are not required for the project. Thus, they have been removed to maintain a clean codebase.

Changed the map as the other map that was destroyed during an accidental rebasing of the main branch.

Changed player movement
so that the map will move instead of the player.
This way, the map will be explorable.